Let FindPatriotSerial func deal with mult sensors
fgetl has been replaced with fread to ensure everything is read from the
serial port.

Length checking has been changed to account for different source sizes.

The response to the 'p' command is converted to a string before being
displayed (required for the binary output from fread).

A "sensors" output is created which, depending on the returned string
length, is either 1 or 2.