Add nomenclature to instructions displayed on GUI
Now reads

Nomenclature: "Locations" are sets of names given to measurements.
"Coordinates" are the X, Y and Z coordinates measured when the
stylus button is pressed. "Data" is everything contained in the table.

For accurate measurements: Ensure the entire stylus is in the northern
measurement hemisphere (the source wire points to the south pole).